Finalize leiden and bppinmf integration

+++ R/clustering.R
@@ -25,7 +25,7 @@
 #' @param nRandomStarts Integer number of random starts. Will pick the
 #' membership with highest quality to return. Default \code{10}.
 #' @param nIterations Integer, maximal number of iterations per random start.
-#' Default \code{100}.
+#' Default \code{5}.
 #' @param useDims Indices of factors to use for clustering. Default \code{NULL}
 #' uses all available factors.
 #' @param groupSingletons Whether to group single cells that make up their own
@@ -37,127 +37,37 @@
 #' @param seed Seed of the random number generator. Default \code{1}.
 #' @param verbose Logical. Whether to show information of the progress. Default
 #' \code{getOption("ligerVerbose")} which is \code{TRUE} if users have not set.
-#' @param partitionType Choose from \code{"ModularityVertexPartition",
-#' "RBConfigurationVertexPartition", "RBERVertexPartition",
-#' "SignificanceVertexPartition", "CPMVertexPartition",
-#' "SurpriseVertexPartition"}. See
-#' \code{\link[leidenbase]{leiden_find_partition}} for detail. Default
-#' \code{"ModularityVertexPartition"}.
-#' @param ... Additional arguments passed to
-#' \code{\link[leidenbase]{leiden_find_partition}}, including
-#' \code{initial_membership} and \code{node_sizes}.
 #' @return \code{object} with refined cluster assignment updated in
 #' \code{clusterName} variable in \code{cellMeta} slot. Can be fetched
 #' with \code{object[[clusterName]]}
-#' @rdname clustering
+#' @rdname runCluster
 #' @export
 #' @examples
-#' pbmcPlot <- runLeidenCluster(pbmcPlot)
+#' pbmcPlot <- runCluster(pbmcPlot)
 #' head(pbmcPlot$leiden_cluster)
-#' pbmcPlot <- runLouvainCluster(pbmcPlot)
+#' pbmcPlot <- runCluster(pbmcPlot, method = "louvain")
 #' head(pbmcPlot$louvain_cluster)
-runLeidenCluster <- function(
-        object,
-        resolution = 1.0,
-        nNeighbors = 20,
-        partitionType = c("ModularityVertexPartition",
-                          "RBConfigurationVertexPartition",
-                          "RBERVertexPartition",
-                          "SignificanceVertexPartition",
-                          "CPMVertexPartition",
-                          "SurpriseVertexPartition"),
-        prune = 1 / 15,
-        eps = 0.1,
-        nRandomStarts = 10,
-        nIterations = 100,
-        useDims = NULL,
-        groupSingletons = TRUE,
-        clusterName = "leiden_cluster",
-        seed = 1,
-        verbose = getOption("ligerVerbose"),
-        ...
-) {
-    partitionType <- match.arg(partitionType)
-    object <- recordCommand(object,
-                            dependencies = c("RANN", "leidenbase", "igraph"))
-    H.norm <- getMatrix(object, "H.norm")
-    if (is.null(H.norm)) {
-        type <- " unnormalized "
-        H.norm <- Reduce(cbind, getMatrix(object, "H"))
-    } else type <- " quantile normalized "
-    if (is.null(H.norm))
-        stop("No factor loading ('H.norm' or 'H') found in `object`.")
-    if (type == " unnormalized ") H.norm <- t(H.norm)
-
-    if (!is.null(useDims)) H.norm <- H.norm[, useDims]
-
-    if (isTRUE(verbose))
-        .log("Leiden clustering on", type, "cell factor loadings...")
-    knn <- RANN::nn2(H.norm, k = nNeighbors, eps = eps)
-    snn <- ComputeSNN(knn$nn.idx, prune = prune)
-    snnSummary <- summary(snn)
-    edges <- as.vector(t(snnSummary[,c(1, 2)]))
-    g <- igraph::graph(edges = edges, n = nrow(snn),directed = FALSE)
-    igraph::E(g)$weight <- snnSummary[, 3]
-
-    set.seed(seed)
-    maxQuality <- -1
-    if (isTRUE(verbose))
-        pb <- utils::txtProgressBar(0, nRandomStarts, style = 3)
-    for (i in seq(nRandomStarts)) {
-        seed <- sample(1000, 1)
-        part <- leidenbase::leiden_find_partition(
-            igraph = g,
-            partition_type = partitionType,
-            resolution_parameter = resolution,
-            edge_weights = igraph::E(g)$weight,
-            num_iter = nIterations,
-            verbose = FALSE,
-            seed = seed,
-            ...
-        )
-        #if (is.null(part$quality)) {
-        #    clusts <- part$membership
-        #    if (isTRUE(verbose)) utils::setTxtProgressBar(pb, nRandomStarts)
-        #    break
-        #}
-        if (part$quality > maxQuality) {
-            clusts <- part$membership
-            maxQuality <- part$quality
-        }
-        if (isTRUE(verbose)) utils::setTxtProgressBar(pb, i)
-    }
-    if (isTRUE(verbose)) cat("\n")
-    names(clusts) <- colnames(object)
-    rownames(snn) <- colnames(object)
-    colnames(snn) <- colnames(object)
-
-    clusts <- groupSingletons(ids = clusts, SNN = snn,
-                              groupSingletons = groupSingletons,
-                              verbose = verbose)
-    if (is.numeric(clusts)) clusts <- clusts - 1
-    clusts <- factor(clusts)
-    cellMeta(object, clusterName, check = FALSE) <- clusts
-    return(object)
-}
-
-#' @rdname clustering
-#' @export
-runLouvainCluster <- function(
+runCluster <- function(
         object,
         resolution = 1.0,
         nNeighbors = 20,
         prune = 1 / 15,
         eps = 0.1,
         nRandomStarts = 10,
-        nIterations = 100,
+        nIterations = 5,
+        method = c("leiden", "louvain"),
         useDims = NULL,
         groupSingletons = TRUE,
-        clusterName = "louvain_cluster",
+        clusterName = paste0(method, "_cluster"),
         seed = 1,
         verbose = getOption("ligerVerbose")
 ) {
-    object <- recordCommand(object, dependencies = "RANN")
+    method <- match.arg(method)
+    object <- switch(method,
+        leiden = recordCommand(object, dependencies = c("RANN", "leidenAlg")),
+        louvain = recordCommand(object, dependencies = c("RANN"))
+    )
+
     H.norm <- getMatrix(object, "H.norm")
     if (is.null(H.norm)) {
         type <- " unnormalized "
@@ -165,40 +75,52 @@ runLouvainCluster <- function(
     } else type <- " quantile normalized "
     if (is.null(H.norm))
         stop("No factor loading ('H.norm' or 'H') found in `object`.")
-    # Not transposing when cbind'ing becausing `t(NULL)` causes error
     if (type == " unnormalized ") H.norm <- t(H.norm)
-
-    edgeOutPath <- paste0("edge_", sub("\\s", "_", Sys.time()), '.txt')
-    edgeOutPath <- gsub("-", "", edgeOutPath)
-    edgeOutPath <- gsub(":", "", edgeOutPath)
 
     if (!is.null(useDims)) H.norm <- H.norm[, useDims]
 
     if (isTRUE(verbose))
-        .log("Louvain clustering on", type, "cell factor loadings...")
+        .log(method, " clustering on", type, "cell factor loadings...")
     knn <- RANN::nn2(H.norm, k = nNeighbors, eps = eps)
     snn <- ComputeSNN(knn$nn.idx, prune = prune)
-    WriteEdgeFile(snn, edgeOutPath, display_progress = FALSE)
-    clusts <- RunModularityClusteringCpp(
-        snn,
-        modularityFunction = 1,
-        resolution = resolution,
-        nRandomStarts = nRandomStarts,
-        nIterations = nIterations,
-        algorithm = 1,
-        randomSeed = seed,
-        printOutput = TRUE,
-        edgefilename = edgeOutPath
-    )
+    if (!is.null(seed)) set.seed(seed)
+    if (method == "leiden") {
+        snnSummary <- summary(snn)
+        edgelist <- as.vector(t(snnSummary[, c(2, 1)])) - 1
+        edgelist_length <- length(edgelist)
+        edge_weights <- snnSummary[,3]
+        clusts <- leidenAlg::find_partition_with_rep_rcpp(
+            edgelist = edgelist, edgelist_length = edgelist_length,
+            num_vertices = nrow(H.norm), direction = FALSE,
+            edge_weights = edge_weights, resolution = resolution,
+            niter = nIterations, nrep = nRandomStarts
+        )
+    } else {
+        edgeOutPath <- paste0("edge_", sub("\\s", "_", Sys.time()), '.txt')
+        edgeOutPath <- gsub("-", "", edgeOutPath)
+        edgeOutPath <- gsub(":", "", edgeOutPath)
+        WriteEdgeFile(snn, edgeOutPath, display_progress = FALSE)
+        clusts <- RunModularityClusteringCpp(
+            snn,
+            modularityFunction = 1,
+            resolution = resolution,
+            nRandomStarts = nRandomStarts,
+            nIterations = nIterations,
+            algorithm = 1,
+            randomSeed = seed,
+            printOutput = TRUE,
+            edgefilename = edgeOutPath
+        )
+        unlink(edgeOutPath)
+    }
+    clusts <- .labelClustBySize(clusts)
     names(clusts) <- colnames(object)
     rownames(snn) <- colnames(object)
     colnames(snn) <- colnames(object)
-    # clusts must not be a factor at this point
     clusts <- groupSingletons(ids = clusts, SNN = snn,
                               groupSingletons = groupSingletons,
                               verbose = verbose)
-    cellMeta(object, columns = clusterName, check = FALSE) <- factor(clusts)
-    unlink(edgeOutPath)
+    cellMeta(object, clusterName, check = FALSE) <- clusts
     return(object)
 }
 
@@ -250,11 +172,11 @@ louvainCluster <- function(
         dims.use = NULL
 ) {
     lifecycle::deprecate_warn("1.99.0", "louvainCluster()",
-                              "runLouvainCluster()")
-    runLouvainCluster(
-        object, resolution = resolution, nNeighbors = k, prune = prune,
-        eps = eps, nRandomStarts = nRandomStarts, nIterations = nIterations,
-        useDims = dims.use, groupSingletons = TRUE,
+                              "runCluster(method = \"louvain\")")
+    runCluster(
+        object, method = "louvain", resolution = resolution, nNeighbors = k,
+        prune = prune, eps = eps, nRandomStarts = nRandomStarts,
+        nIterations = nIterations, useDims = dims.use, groupSingletons = TRUE,
         clusterName = "louvain_cluster", seed = random.seed, verbose = verbose
     )
 }
@@ -309,4 +231,15 @@ groupSingletons <- function(
         .log(length(singletons), " singletons identified. ",
              length(unique(ids)), " final clusters.")
     return(ids)
+}
+
+.labelClustBySize <- function(clusts) {
+    clusts <- as.character(clusts)
+    count <- data.frame(table(clusts))
+    count <- count[order(count$Freq, decreasing = TRUE),]
+    map <- seq(0, nrow(count) - 1)
+    names(map) <- count[[1]]
+    newClusts <- map[clusts]
+    newClusts <- factor(newClusts)
+    return(newClusts)
 }
